command_center/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
google-adk>=0.1.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.28.0
google-genai>=0.2.0
httpx>=0.25.0